requirements.txt
pymongo==4.6.1
pandas==2.2.0
python-dotenv==1.0.1
jupyter==1.0.0
notebook==7.0.7 